A few portability fixes from Martin back on Aug 27/97.
Submitted by:	Martin Kraemer
Reviewed by:	Dean Gaudet


git-svn-id: https://svn.apache.org/repos/asf/httpd/httpd/branches/1.3.x@79263 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/util_script.c
@@ -499,7 +499,8 @@ API_EXPORT(void) send_size(size_t size, request_rec *r)
 {
     char ss[20];
 
-    if (size == -1)
+    /* XXX: this -1 thing is a gross hack */
+    if (size == (size_t)-1)
 	strcpy(ss, "    -");
     else if (!size)
 	strcpy(ss, "   0k");
